Add files via upload
This is a C program where an electronic concept has been converted into a program. The 220V AC will be rectified to 15V DC, and from there, the required voltage will be supplied to charge the device as needed. The entire process will be handled automatically by the program—you only need to specify which device you want to charge.